requirements.txt
# Core dependencies
pyyaml>=6.0
python-dotenv>=1.0.0
schedule>=1.2.0
requests>=2.28.0
pillow>=9.0.0

# Database
sqlite3>=0.0.0

# AI Models (uncomment as needed)
# llama-cpp-python>=0.2.0  # For Llama models
# diffusers>=0.14.0        # For Stable Diffusion
# transformers>=4.28.0     # For transformer-based models
# torch>=2.0.0             # PyTorch for AI models
# accelerate>=0.19.0       # For accelerated inference

# Social Media APIs (uncomment as needed)
# tweepy>=4.14.0           # Twitter API
# instagrapi>=1.16.30      # Instagram API
# facebook-sdk>=3.1.0      # Facebook API

# Utilities
tqdm>=4.65.0
python-dateutil>=2.8.2